# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kidspeak Language Learning Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - in: query
          name: month
          schema:
            type: string
            description: "YYYY-MM format"
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
          enum: [rent, utilities, salaries, materials, maintenance, other]
        description:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - description
        - amount
        - expenseDate
        - createdAt
    CreateExpenseBody:
      type: object
      properties:
        category:
          type: string
          enum: [rent, utilities, salaries, materials, maintenance, other]
        description:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      required:
        - category
        - description
        - amount
        - expenseDate
...